+++ CHANGELOG.md
@@ -9,6 +9,7 @@ CHANGELOG
 
 ### Frontend
 - Show version including link to upstream in footer (#49).
+- Remove inout field text, does not work anyway.
 
 0.1.0.alpha2
 ------------