+++ CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
 ### Backend
 - Data in the columns assigned to the `extra` field will be saved as `{"data0": ..., "data1": ..., ...}` (#35).
 - Constant fields can be configured with the configuration parameter `"constant_fields"` (#38).
+- Additional custom input fields can be added with the configuration parameter `"custom_input_fields"` (#48).
 
 ### Frontend
 - Show version including link to upstream in footer (#49).